# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Obiara API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Authoritative REST contract for Obiara member clients. Durable API changes
    are additive within v1. Retriable writes require an Idempotency-Key.
...
servers:
  - url: /
    description: API root
...
paths:
...
  /v1/admin/verifications:
    get:
      operationId: listAdminVerificationQueue
      tags: [Admin Verification]
      summary: List privacy-redacted verification cases
      description: Requires the verification.queue.read capability.
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Oldest queued verification cases, without raw evidence.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVerificationQueueEnvelope"
        "403":
          $ref: "#/components/responses/AdminForbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/verifications/{id}:
    get:
      operationId: getAdminVerificationCase
      tags: [Admin Verification]
      summary: Read privacy-redacted verification case detail
      description: Requires the verification.queue.read capability.
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Redacted case detail; raw card and birth date are excluded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVerificationCaseEnvelope"
        "403":
          $ref: "#/components/responses/AdminForbidden"
        "404":
          $ref: "#/components/responses/VerificationCaseNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/verifications/{id}/evidence-access:
    post:
      operationId: accessAdminVerificationEvidence
      tags: [Admin Verification]
      summary: Access bounded verification evidence and create an audit event
      description: |
        Requires verification.evidence.read, recent MFA, a bounded purpose,
        and a reason. Every successful access writes a dedicated audit event.
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceAccessInput"
      responses:
        "200":
          description: Bounded evidence returned after the audit write succeeds.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationEvidenceEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "403":
          $ref: "#/components/responses/AdminForbidden"
        "404":
          $ref: "#/components/responses/VerificationCaseNotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/verifications/{id}/decisions:
    post:
      operationId: decideAdminVerificationCase
      tags: [Admin Verification]
      summary: Approve or reject a queued verification case
      description: |
        Requires verification.review. The reason, expected version, and
        Idempotency-Key make retries deterministic and stale writes fail closed.
      security:
        - AdminBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationDecisionInput"
      responses:
        "200":
          description: Decision committed or an identical command replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationDecisionEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "403":
          $ref: "#/components/responses/AdminForbidden"
        "404":
          $ref: "#/components/responses/VerificationCaseNotFound"
        "409":
          $ref: "#/components/responses/VerificationCaseConflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/circles/{circleId}/ampe:
    post:
      operationId: createAmpeRound
      tags: [Community]
      summary: Start a private Ampe round in an exact two-member circle
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "201":
          description: Private round created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmpeEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AmpeNotAvailable"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/ampe/{roundId}:
    get:
      operationId: getAmpeRound
      tags: [Community]
      summary: Read a privacy-safe Ampe round projection
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/AmpeRoundId"
      responses:
        "200":
          description: Current member-relative round state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmpeEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AmpeNotAvailable"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/circles/{circleId}/ampe/{roundId}/commands:
    post:
      operationId: commandAmpeRound
      tags: [Community]
      summary: Mark ready or lock a manual private gesture
      security:
        - MemberBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CircleRoomId"
        - $ref: "#/components/parameters/AmpeRoundId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmpeCommandInput"
      responses:
        "200":
          description: Command accepted and current projection returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmpeEnvelope"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/AmpeNotAvailable"
        "409":
          $ref: "#/components/responses/AmpeConflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    MemberBearer:
      type: http
      scheme: bearer
      bearerFormat: Opaque
      description: Short-lived member access token; the session subject must match memberId.
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Staff token carrying the explicitly documented capability and recent-MFA claim where required.
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Safe caller-provided identifier; invalid values are replaced.
      schema:
        $ref: "#/components/schemas/CorrelationId"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable key reused for retries of the same command.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
...
    CircleRoomId:
      name: circleId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    AmpeRoundId:
      name: roundId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  headers:
    CorrelationId:
      required: true
      description: Effective request correlation identifier.
      schema:
        $ref: "#/components/schemas/CorrelationId"
  responses:
    Unauthorized:
      description: A valid authenticated session is required.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AmpeNotAvailable:
      description: The round or its exact two-member circle context is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AmpeConflict:
      description: The round sequence changed before the command was applied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InvalidJSON:
      description: The body is malformed, oversized, contains unknown fields, or contains multiple values.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidJSON:
              value:
                error:
                  code: invalid_json
                  message: The request body must be one valid JSON object.
                meta:
                  correlationId: 01J9EXAMPLE000000000000000
    UnsupportedMediaType:
      description: Content-Type is not application/json.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AdminForbidden:
      description: The staff principal lacks the exact capability, or recent MFA where evidence access requires it.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    VerificationCaseNotFound:
      description: No verification case with this identifier exists.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    VerificationCaseConflict:
      description: The case is closed, stale, or the idempotency key belongs to another decision.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationFailed:
      description: Request fields or the idempotency key are invalid.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidEmail:
              value:
                error:
                  code: validation_failed
                  message: One or more fields are invalid.
                  details:
                    - field: email
                      reason: must be a valid email address
                meta:
                  correlationId: 01J9EXAMPLE000000000000000
    InternalError:
      description: An unexpected server failure occurred.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    CorrelationId:
      type: string
      minLength: 8
      maxLength: 128
      pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
...
    Metadata:
      type: object
      additionalProperties: false
      required: [correlationId]
      properties:
        correlationId:
          $ref: "#/components/schemas/CorrelationId"
...
    AdminVerificationCaseData:
      type: object
      additionalProperties: false
      required: [caseId, subjectRef, reasonCode, submittedAt, version]
      properties:
        caseId:
          type: string
        subjectRef:
          type: string
          description: Stable pseudonymous subject reference; never the account identifier.
        reasonCode:
          type: string
          enum: [provider_outage, provider_uncertain, manual_review]
        submittedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [queued_manual, approved, rejected]
        version:
          type: integer
          format: int64
          minimum: 1
    AdminVerificationCaseEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AdminVerificationCaseData"
        meta:
          $ref: "#/components/schemas/Metadata"
    AdminVerificationQueueData:
      type: object
      additionalProperties: false
      required: [cases]
      properties:
        cases:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/AdminVerificationCaseData"
    AdminVerificationQueueEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AdminVerificationQueueData"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    EvidenceAccessInput:
      type: object
      additionalProperties: false
      required: [purpose, reason]
      properties:
        purpose:
          type: string
          minLength: 3
          maxLength: 64
          pattern: ^[a-z0-9][a-z0-9_.-]{2,63}$
        reason:
          type: string
          minLength: 8
          maxLength: 240
    VerificationEvidenceData:
      type: object
      additionalProperties: false
      required: [caseId, maskedCard, ageBand, providerStatus]
      properties:
        caseId:
          type: string
        maskedCard:
          type: string
          description: Only the final four card characters are visible.
        ageBand:
          type: string
          enum: [under_18, 18_24, 25_34, 35_49, 50_plus]
        providerStatus:
          type: string
          enum: [provider_outage, provider_uncertain, manual_review]
    VerificationEvidenceEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/VerificationEvidenceData"
        meta:
          $ref: "#/components/schemas/Metadata"
    VerificationDecisionInput:
      type: object
      additionalProperties: false
      required: [outcome, reason, expectedVersion]
      properties:
        outcome:
          type: string
          enum: [approve, reject]
        reason:
          type: string
          minLength: 8
          maxLength: 240
        expectedVersion:
          type: integer
          format: int64
          minimum: 1
    VerificationDecisionData:
      type: object
      additionalProperties: false
      required: [case, outcome, replayed]
      properties:
        case:
          $ref: "#/components/schemas/AdminVerificationCaseData"
        outcome:
          type: string
          enum: [approve, reject]
        replayed:
          type: boolean
    VerificationDecisionEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/VerificationDecisionData"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    AmpeCommandInput:
      type: object
      additionalProperties: false
      required: [action, expectedSequence]
      properties:
        action:
          type: string
          enum: [ready, lock]
        choice:
          type: string
          enum: [together, apart]
        expectedSequence:
          type: integer
          format: int64
          minimum: 0
    AmpePlayerData:
      type: object
      additionalProperties: false
      required: [ready, connected, locked]
      properties:
        ready:
          type: boolean
        connected:
          type: boolean
        locked:
          type: boolean
    AmpeData:
      type: object
      additionalProperties: false
      required: [id, sequence, you, other, paused, complete]
      properties:
        id:
          type: string
        sequence:
          type: integer
          format: int64
          minimum: 0
        you:
          $ref: "#/components/schemas/AmpePlayerData"
        other:
          $ref: "#/components/schemas/AmpePlayerData"
        paused:
          type: boolean
        ownChoice:
          type: string
          enum: [together, apart]
        yourReveal:
          type: string
          enum: [together, apart]
        otherReveal:
          type: string
          enum: [together, apart]
        complete:
          type: boolean
    AmpeEnvelope:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AmpeData"
        meta:
          $ref: "#/components/schemas/Metadata"
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          pattern: ^[a-z][a-z0-9_]*$
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        meta:
          $ref: "#/components/schemas/Metadata"
...